% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correction.r
\name{biasCorrection.jmleIRT}
\alias{biasCorrection.jmleIRT}
\title{Analytical bias correction for jmleIRT objects}
\usage{
\method{biasCorrection}{jmleIRT}(jmle_obj, theta = NULL, beta = NULL, X = NULL, I = NULL, ...)
}
\arguments{
\item{jmle_obj}{An object of class \code{"jmleIRT"} as returned by}

\item{theta}{Person parameters (ignored if jmle_obj is provided)}

\item{beta}{Item parameters (ignored if jmle_obj is provided)}

\item{X}{Response matrix (ignored if jmle_obj is provided)}

\item{I}{Number of items (ignored if jmle_obj is provided)
\code{jmle_estimation()}.}

\item{...}{further arguments, but currently unused.}
}
\value{
An object of class \code{"biasCorrection"} containing numeric
vectors \code{theta} and \code{beta}.
}
\description{
Applies the C++ first-order bias correction to an object of class
\code{jmleIRT} and returns corrected person and item parameters.
}
