% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_jmleIRT.r
\name{coef.jmleIRT}
\alias{coef.jmleIRT}
\title{Coefficients for jmleIRT objects}
\usage{
\method{coef}{jmleIRT}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"jmleIRT"}.}

\item{...}{Currently unused.}
}
\value{
A list with components \code{theta} and \code{beta}.
}
\description{
Coefficients for jmleIRT objects
}
