% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmlIRT-package.r
\docType{package}
\name{jmleIRT-package}
\alias{jmleIRT}
\alias{jmleIRT-package}
\title{jmleIRT: Joint Maximum Likelihood Estimation of the Rasch Model}
\description{
Implements joint maximum likelihood estimation (JMLE) for item parameters in the Rasch model (1PL). Provides methods for bias correction of item parameter estimates and weighted likelihood estimation (WLE) of person parameters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jansteinfeld.github.io/jmleIRT/}
  \item \url{https://github.com/jansteinfeld/jmleIRT}
  \item Report bugs at \url{https://github.com/jansteinfeld/jmleIRT/issues}
}

}
\author{
\strong{Maintainer}: Jan Steinfeld \email{jan.d.steinfeld@gmail.com} (\href{https://orcid.org/0000-0001-9853-8260}{ORCID})

Authors:
\itemize{
  \item Alexander Robitzsch \email{robitzsch@leibniz-ipn.de} (\href{https://orcid.org/0000-0002-8226-3132}{ORCID})
}

}
\keyword{internal}
