% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik_jmleIRT.r
\name{logLik.jmleIRT}
\alias{logLik.jmleIRT}
\title{Log-likelihood for jmleIRT objects}
\usage{
\method{logLik}{jmleIRT}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"jmleIRT"}.}

\item{...}{Currently unused.}
}
\value{
An object of class \code{"logLik"}.
}
\description{
Computes the joint log-likelihood of the Rasch model at the JMLE estimates.
}
