% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_jmleIRT.r
\name{print.jmleIRT}
\alias{print.jmleIRT}
\title{Print method for JMLE Rasch Model Objects}
\usage{
\method{print}{jmleIRT}(x, ...)
}
\arguments{
\item{x}{An object of class \code{jmleIRT} returned from a JMLE Rasch model estimation function.}

\item{...}{Additional arguments passed to or from other methods (currently ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
This function prints a brief summary of a Joint Maximum Likelihood Estimation (JMLE)
Rasch model object of class \code{jmleIRT}.
}
\details{
The print method displays basic information including:
- The type of the model object
- Number of iterations used in the estimation procedure
- Prompts the user to use \code{summary()} for more detailed output.
}
