% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_jmleIRT.r
\name{summary.biasCorrection}
\alias{summary.biasCorrection}
\title{Summary for Bias Corrected JMLE Estimates}
\usage{
\method{summary}{biasCorrection}(object, ...)
}
\arguments{
\item{object}{Object of class \code{"biasCorrection"} containing \code{beta} and \code{theta}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input \code{object}, invisibly.
}
\description{
Print a summary of bias-corrected item difficulty and person ability estimates.
}
